Fixed backward compatibility for patches
Message patches that were previously stored on the database (or received by a peer) would break because we changed the bincode representation.
Now the message patch was split into two. The first one is the same message patch when icechat did not support attachments, the new one is a patch that includes attachment information.